feat(scanner): comment, grouping stuff and operators

+++ logger/include/logger.h
@@ -10,7 +10,7 @@ class logger final
 		: public base::singleton<logger>
 {
 public:
-	void error(int line,const std::string& message);
+	void error(size_t line,const std::string& message);
 
 private:
 	size_t errors_{0};
