+++ logger/logger.cpp
@@ -5,7 +5,7 @@
 
 using namespace std;
 
-void clox::logger::logger::error(int line, const std::string& message)
+void clox::logger::logger::error(size_t line, const std::string& message)
 {
 	errors_++;
 	cout << format("[Line {}] Error: {}", line, message) << endl;
